+++ src/obr/init.ts
@@ -68,9 +68,13 @@ const initScene = async () => {
             hpBarOffset: 0,
             acOffset: { x: 0, y: 0 },
             acShield: true,
+            ruleset: "e5",
             allowNegativeNumbers: false,
             id: uuidv4(),
             groups: ["Default"],
+            openGroups: ["Default"],
+            initiativeDice: 20,
+            statblockPopover: { width: 500, height: 600 },
             playerSort: false,
         };
     } else {
@@ -86,6 +90,9 @@ const initScene = async () => {
         }
         if (!sceneData.groups || sceneData.groups.length === 0) {
             sceneData.groups = ["Default"];
+        }
+        if (sceneData.openGroups === undefined) {
+            sceneData.openGroups = ["Default"];
         }
         if (sceneData.hpBarSegments === undefined) {
             sceneData.hpBarSegments = 0;
@@ -109,6 +116,12 @@ const initScene = async () => {
         }
         if (!sceneData.acShield) {
             sceneData.acShield = true;
+        }
+        if (!sceneData.initiativeDice) {
+            sceneData.initiativeDice = 20;
+        }
+        if (sceneData.statblockPopover === undefined) {
+            sceneData.statblockPopover = { width: 500, height: 600 };
         }
         if (!sceneData.playerSort) {
             sceneData.playerSort = false;